Optimize parser memory consumption. Remove 'prev' and 'block_size' fields form the linked list structure. Remove 'magic' field from structures used in parser.

+++ jerry-core/parser/collections/linked-list.h
@@ -23,7 +23,7 @@ typedef uint8_t* linked_list;
 
 linked_list linked_list_init (uint16_t);
 void linked_list_free (linked_list);
-void *linked_list_element (linked_list, uint16_t);
-void linked_list_set_element (linked_list, uint16_t, void *);
+void *linked_list_element (linked_list, size_t);
+void linked_list_set_element (linked_list, size_t, void *);
 
 #endif /* LINKED_LIST_H */
